requirements.txt
blis==0.7.5
catalogue==2.0.6
certifi==2021.10.8
charset-normalizer==2.0.8
click==8.0.3
cymem==2.0.6
DAWG-Python==0.7.2
docopt==0.6.2
en-core-web-sm==3.2.0
Flask==2.0.2
Flask-Cors==3.0.10
Flask-Login==0.5.0
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
JSON-log-formatter==0.4.0
langcodes==3.3.0
MarkupSafe==2.0.1
murmurhash==1.0.6
nest-asyncio==1.5.4
numpy==1.21.4
oauthlib==3.1.1
packaging==21.3
pandas==1.3.4
pathy==0.6.1
preshed==3.0.6
pydantic==1.8.2
pymorphy2==0.9.1
pymorphy2-dicts-ru==2.4.417127.4579844
pyparsing==3.0.6
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
requests==2.26.0
requests-oauthlib==1.3.0
six==1.16.0
smart-open==5.2.1
spacy==3.2.0
spacy-legacy==3.0.8
spacy-loggers==1.0.1
srsly==2.4.2
thinc==8.0.13
tqdm==4.62.3
tweepy==4.4.0
typer==0.4.0
typing-extensions==4.0.1
urllib3==1.26.7
wasabi==0.8.2
Werkzeug==2.0.2
